Complete Task 1

+++ src/DataTypes.java
@@ -1,7 +1,8 @@
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(int[] numbers) {
-        int s = 0; // variable to accumulate the sum in!
+        // the original code didn't pass because the test case uses a number that is larger than 32-bit int can handle
+        // this causes an overflow so the result is less than the actual sum
+        long s = 0; // variable to accumulate the sum in!
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
